Clean up builtin functions
